fix(service_state): handle preempt during service call
If preemption is requested while calling service the preemption is not handled and the preempt requested flag is left in an active state. The next time the state is activated it will immediately service the preemption and return 'preempted'. This fix checks for preemption before processing the result and preempts if requested.