fix: wait for server for every execution
Wait for action server to be ready for every execution of the state. This ensures that the action server is still running at execution. Earlier implementation would hang forever if the action server died at some point between initialization of the state and execution (any execution) of the state.

In the case of a missing action server a new outcome "unreachable" is returned. This allows the user of this state to detect that the action server no longer is running.